# server/v2/dataservice/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Scholar DataService API",
    "description": "V2 Sprint-2 DataService API for Scholar AI Advisor ecosystem. Provides secure, audited access to scholar data with FERPA compliance and role-based access control.",
    "version": "2.0.0",
    "contact": {
      "name": "Scholar AI Advisor Team",
      "url": "https://scholaraiadvisor.com"
    }
  },
  "servers": [
    {
      "url": "/api/v2/dataservice",
      "description": "DataService API"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "summary": "List Users",
        "description": "Retrieve paginated list of users. FERPA-sensitive fields are redacted for consumer role.",
        "operationId": "listUsers",
        "tags": ["Users"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/ViewerRole" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create User",
        "description": "Create a new user record",
        "operationId": "createUser",
        "tags": ["Users"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "summary": "Get User",
        "description": "Retrieve a single user by ID. FERPA-sensitive fields are redacted for consumer role.",
        "operationId": "getUser",
        "tags": ["Users"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/ViewerRole" },
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Update User",
        "description": "Update an existing user record",
        "operationId": "updateUser",
        "tags": ["Users"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete User",
        "description": "Delete a user record",
        "operationId": "deleteUser",
        "tags": ["Users"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "responses": {
          "204": { "description": "User deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/providers": {
      "get": {
        "summary": "List Providers",
        "description": "Retrieve paginated list of providers (B2B organizations)",
        "operationId": "listProviders",
        "tags": ["Providers"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" },
          {
            "name": "is_active",
            "in": "query",
            "schema": { "type": "boolean", "default": true }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create Provider",
        "description": "Create a new provider (B2B organization)",
        "operationId": "createProvider",
        "tags": ["Providers"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProviderRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/providers/{id}": {
      "get": {
        "summary": "Get Provider",
        "description": "Retrieve a single provider by ID",
        "operationId": "getProvider",
        "tags": ["Providers"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Update Provider",
        "description": "Update an existing provider",
        "operationId": "updateProvider",
        "tags": ["Providers"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProviderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete Provider",
        "description": "Delete a provider record",
        "operationId": "deleteProvider",
        "tags": ["Providers"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/ResourceId" }
        ],
        "responses": {
          "204": { "description": "Provider deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/scholarships": {
      "get": {
        "summary": "List Scholarships",
        "description": "Retrieve paginated list of active scholarships",
        "operationId": "listScholarships",
        "tags": ["Scholarships"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScholarshipListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create Scholarship",
        "description": "Create a new scholarship record",
        "operationId": "createScholarship",
        "tags": ["Scholarships"],
        "parameters": [
          { "$ref": "#/components/parameters/TraceId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateScholarshipRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scholarship created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScholarshipResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "parameters": {
      "TraceId": {
        "name": "X-Trace-Id",
        "in": "header",
        "required": true,
        "description": "Unique trace identifier for request tracking",
        "schema": { "type": "string" }
      },
...
      "ViewerRole": {
        "name": "X-Viewer-Role",
        "in": "header",
        "required": false,
        "description": "Viewer role for FERPA field filtering",
        "schema": {
          "type": "string",
          "enum": ["consumer", "school_official", "admin", "system"],
          "default": "consumer"
        }
      },
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "default": 100, "maximum": 1000 }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": { "type": "integer", "default": 0 }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "schemas": {
...
      "UserListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "meta": {
            "type": "object",
            "properties": {
              "limit": { "type": "integer" },
              "offset": { "type": "integer" },
              "total": { "type": "integer" }
            }
          }
        }
      },
...
    }
  },
...
}